PNC exceptions can have duplicate paths on the AHO XML
There is a priority when generating the XML where it will show the exception
on the ASN element if no other exception exists
